requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
faiss-cpu
pypdf
huggingface-hub
flask
python-dotenv
sentence-transformers
